backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
httpx>=0.27
supabase>=2.4
python-dotenv>=1.0
pydantic>=2.6
